implement interface Inventory; small refactor;
