+++ src/test/java/com/gebarowski/tad_inventory/InventoryImplTest.java
@@ -1,0 +1,10 @@
+package com.gebarowski.tad_inventory;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ * Created by Michał Gębarowski on 04/04/2018
+ */
+class InventoryImplTest {
+
+}
